demo login

+++ app/Http/Controllers/UserController.php
@@ -6,21 +6,28 @@ use App\Http\Requests\StoreUserRequest;
 use App\Models\Role;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
     function create() {
+        if (!$this->userCan('crud-user')) {
+            abort(403);
+        }
         $roles = Role::all();
         return view('admin.users.create', compact('roles'));
     }
 
     function store(StoreUserRequest $request) {
+        if (!$this->userCan('crud-user')) {
+            abort(403);
+        }
         // tao user
         $user = new User();
         $user->name = $request->name;
         $user->email = $request->email;
-        $user->password = Hash::make($request->email);
+        $user->password = Hash::make($request->password);
         $user->save();
 
         //them role user
@@ -36,6 +43,9 @@ class UserController extends Controller
     }
 
     function edit(Request $request, $id) {
+        if (!Gate::allows('crud-user')) {
+            abort(403);
+        }
         $user = User::findOrFail($id);
         $user->name = $request->name;
         $user->save();
@@ -44,10 +54,12 @@ class UserController extends Controller
     }
 
     function delete($id){
+        if (!Gate::allows('crud-user')) {
+            abort(403);
+        }
         $user = User::findOrFail($id);
         $user->roles()->detach();
         $user->delete();
-
-        return response()->json(['message' => 'delete successfully!']);
+        return response()->json(['message' => __('message.delete_success')]);
     }
 }
